Allowing users to choice DE test MixScape Uses
Hi,

I've been playing with mixscape on yours and some of my own scRNA-Seq Crispr Screen data and noticed that you don't allow the user to set which differential expression test to use. I think that should be parameter for mixscape as the algorithm relays upon the DEGs it's detected. 

Kind Regards,

Alex Baker